Separate the game model to model.py
